chore: relocate SemanticPreview component
