+++ components/button/demo/_semantic.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { AntDesignOutlined } from '@ant-design/icons';
 import { Button } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/builtins/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
