+++ components/drawer/demo/_semantic.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Drawer, Typography } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/builtins/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 const locales = {
