+++ components/list/demo/_semantic.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
+import SemanticPreview from '../../../.dumi/theme/builtins/SemanticPreview';
 import useLocale from '../../../.dumi/hooks/useLocale';
 
 import { Avatar, List, Space } from 'antd';
